little improve games

+++ app/src/main/java/com/teamx/equiz/data/models/orderDetailData/CartDetail.kt
@@ -3,6 +3,6 @@ package com.teamx.equiz.data.models.orderDetailData
 data class CartDetail(
     val __v: Int,
     val _id: String,
-    val cartPrice: Int,
+    val cartPrice: Double,
     val productDetails: List<ProductDetail>
 )
